platform-dependent-components/problem-solver/py/requirements.txt
py-sc-kpm==0.2.0
g4f==0.2.9.11
py-sc-client==0.3.0
